(android) Fix show_intro flag after legacy migration
Show_intro flag now falls back to legacy prefs when undefined.
This prevents race conditions where the app would show the
intro screens when not needed.

Data migration is also streamlined, prefs and payments are now
migrated in the same routine.
